Converting project back into a binary
