+++ data/preprocessor_out.txt
@@ -1,0 +1,19 @@
+very
+GLOBAL_0
+STR_0
+very
+GLOBAL_1
+FUNC_START_0
+plz
+GLOBAL_1
+STR_1
+STR_2
+very
+GLOBAL_2
+ARR_START_0
+very
+GLOBAL_3
+FUNC_START_1
+very
+GLOBAL_4
+ARR_START_1
